+++ packages/common/src/sites/_internal/SiteSchema.ts
@@ -2,7 +2,7 @@ import { IConfigurationSchema } from "../../core";
 import { HubItemEntitySchema } from "../../core/schemas/shared/HubItemEntitySchema";
 
 export type SiteEditorType = (typeof SiteEditorTypes)[number];
-export const SiteEditorTypes = ["hub:site:edit"] as const;
+export const SiteEditorTypes = ["hub:site:edit", "hub:site:create"] as const;
 
 /**
  * defines the JSON schema for a Hub Site's editable fields