+++ packages/common/test/content/_internal/ContentUiSchemaEdit.test.ts
@@ -1,0 +1,157 @@
+import { buildUiSchema } from "../../../src/content/_internal/ContentUiSchemaEdit";
+import { MOCK_CONTEXT } from "../../mocks/mock-auth";
+import * as getCategoryItemsModule from "../../../src/core/schemas/internal/getCategoryItems";
+import * as getLocationExtentModule from "../../../src/core/schemas/internal/getLocationExtent";
+import * as getLocationOptionsModule from "../../../src/core/schemas/internal/getLocationOptions";
+import * as getTagItemsModule from "../../../src/core/schemas/internal/getTagItems";
+
+describe("buildUiSchema: content edit", () => {
+  it("returns the full content edit uiSchema", async () => {
+    spyOn(getCategoryItemsModule, "getCategoryItems").and.returnValue(
+      Promise.resolve([])
+    );
+    spyOn(getLocationExtentModule, "getLocationExtent").and.returnValue(
+      Promise.resolve([])
+    );
+    spyOn(getLocationOptionsModule, "getLocationOptions").and.returnValue(
+      Promise.resolve([])
+    );
+    spyOn(getTagItemsModule, "getTagItems").and.returnValue(
+      Promise.resolve([])
+    );
+
+    const uiSchema = await buildUiSchema(
+      "some.scope",
+      { thumbnailUrl: "https://some-thumbnail-url.com" } as any,
+      MOCK_CONTEXT
+    );
+    expect(uiSchema).toEqual({
+      type: "Layout",
+      elements: [
+        {
+          type: "Section",
+          labelKey: "some.scope.sections.basic.title",
+          elements: [
+            {
+              labelKey: "some.scope.fields.title.label",
+              scope: "/properties/name",
+              type: "Control",
+              options: {
+                helperText: {
+                  labelKey: "some.scope.fields.title.hint",
+                },
+                messages: [
+                  {
+                    type: "ERROR",
+                    keyword: "required",
+                    icon: true,
+                    labelKey: "some.scope.fields.title.requiredError",
+                  },
+                ],
+              },
+            },
+            {
+              labelKey: "some.scope.fields.summary.label",
+              scope: "/properties/summary",
+              type: "Control",
+              options: {
+                control: "hub-field-input-input",
+                type: "textarea",
+                helperText: {
+                  labelKey: "some.scope.fields.summary.hint",
+                },
+              },
+            },
+            {
+              labelKey: "some.scope.fields.description.label",
+              scope: "/properties/description",
+              type: "Control",
+              options: {
+                control: "hub-field-input-input",
+                type: "textarea",
+                helperText: {
+                  labelKey: "some.scope.fields.description.hint",
+                },
+              },
+            },
+            {
+              labelKey: "some.scope.fields._thumbnail.label",
+              scope: "/properties/_thumbnail",
+              type: "Control",
+              options: {
+                control: "hub-field-input-image-picker",
+                imgSrc: "https://some-thumbnail-url.com",
+                maxWidth: 727,
+                maxHeight: 484,
+                aspectRatio: 1.5,
+                helperText: {
+                  labelKey: "some.scope.fields._thumbnail.helperText",
+                },
+              },
+            },
+            {
+              labelKey: "some.scope.fields.tags.label",
+              scope: "/properties/tags",
+              type: "Control",
+              options: {
+                control: "hub-field-input-combobox",
+                items: [],
+                allowCustomValues: true,
+                selectionMode: "multiple",
+                placeholderIcon: "label",
+                helperText: { labelKey: "some.scope.fields.tags.hint" },
+              },
+            },
+            // categories
+            {
+              labelKey: "some.scope.fields.categories.label",
+              scope: "/properties/categories",
+              type: "Control",
+              options: {
+                control: "hub-field-input-combobox",
+                items: [],
+                allowCustomValues: false,
+                selectionMode: "multiple",
+                placeholderIcon: "select-category",
+                helperText: {
+                  labelKey: "some.scope.fields.categories.agolHint", // TODO: hint should describe whether it can be set on Enterprise or Online
+                },
+              },
+            },
+            {
+              labelKey: "some.scope.fields.license.label",
+              scope: "/properties/licenseInfo",
+              type: "Control",
+              options: {
+                control: "arcgis-hub-license-picker",
+                helperText: {
+                  labelKey: "some.scope.fields.license.helperText",
+                },
+              },
+            },
+          ],
+        },
+        {
+          type: "Section",
+          labelKey: "some.scope.sections.location.label",
+          options: {
+            helperText: {
+              labelKey: "some.scope.sections.location.helperText",
+            },
+          },
+          elements: [
+            {
+              scope: "/properties/location",
+              type: "Control",
+              options: {
+                control: "hub-field-input-location-picker",
+                extent: [],
+                options: [],
+              },
+            },
+          ],
+        },
+      ],
+    });
+  });
+});